+++ axum/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `serde_json::Error` ([#1371])
 - **added**: `JsonRejection` now displays the path at which a deserialization
   error occurred too ([#1371])
+- **fixed:** Support streaming/chunked requests in `ContentLengthLimit`
 
 [#1371]: https://github.com/tokio-rs/axum/pull/1371